==> notebooks/4_build_numeric_dataframe.ipynb <==
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd
import warnings
import seaborn as sns

from scipy.stats import pearsonr

sns.set_style("darkgrid")
np.random.seed(930525)
pd.set_option('display.max_columns', 20)
pd.set_option('display.max_rows', 200)

warnings.simplefilter('once')

%matplotlib inline
%load_ext watermark
%watermark --iversions

train = pd.read_csv("../data/train.txt", sep="\t", index_col=0)
train_warp = pd.read_csv("../data/train.warp.txt", sep="\t", index_col=0)

train.shape

train_warp.shape

# filter out columns that are correlated
s_cols = set(["tax", "food", "kegg", "nutrients"])
s_cols_response = set(["day_plus_one"])
s_cont_cols = set(["Age", "Weight", "Height", "BMI", "oilGrams.assigned", "Activity.Factor", "StudyDayNo"])
df_numeric = train[[col for col in train.columns if col.split(".")[0] in s_cols or col in s_cont_cols]]
df_response = train[[col for col in train.columns if col.split(".")[0] in s_cols_response]]
df_binary = train[[col for col in train.columns if (col.split(".")[0] not in s_cols.union(s_cols_response) and col not in s_cont_cols) and col != "Timing"]]

df_binary.to_csv("../data/prediction.train.binary.txt", sep="\t")
df_response.to_csv("../data/prediction.train.response.txt", sep="\t")
df_numeric.to_csv("../data/prediction.train.numeric.txt", sep="\t")

# filter out columns that are correlated
s_cols = set(["tax", "food", "kegg", "nutrients"])
s_cols_response = set(["day_plus_one"])
s_cont_cols = set(["Age", "Weight", "Height", "BMI", "oilGrams.assigned", "Activity.Factor", "StudyDayNo"])
df_numeric_warp = train_warp[[col for col in train_warp.columns if col.split(".")[0] in s_cols or col in s_cont_cols]]
df_response_warp = train_warp[[col for col in train_warp.columns if col.split(".")[0] in s_cols_response]]
df_binary_warp = train_warp[[col for col in train_warp.columns if (col.split(".")[0] not in s_cols.union(s_cols_response) and col not in s_cont_cols) and col != "Timing"]]

df_binary_warp.to_csv("../data/prediction.train.binary.warp.txt", sep="\t")
df_response_warp.to_csv("../data/prediction.train.response.warp.txt", sep="\t")
df_numeric_warp.to_csv("../data/prediction.train.numeric.warp.txt", sep="\t")

df_binary_warp

[_ for _ in df_response_warp.columns if "other" in _]
