==> notebooks/9_export_network_cytoscape.ipynb <==
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd
import warnings
import seaborn as sns

from scipy.stats import pearsonr

sns.set_style("darkgrid")
np.random.seed(930525)
pd.set_option('display.max_columns', 20)
pd.set_option('display.max_rows', 200)

warnings.simplefilter('once')

%matplotlib inline
%load_ext watermark
%watermark --iversions

import networkx as nx
import statsmodels.formula.api as smf

df_dag_train = pd.read_csv("../results/dag.prediction.train.response.updated.csv", sep="\t")
df_dag_train_warp = pd.read_csv("../results/dag.prediction.train.response.warp.updated.csv", sep="\t")
df_train = pd.read_csv("../data/train.txt", index_col=0, header = 0, sep="\t")
df_train_warp = pd.read_csv("../data/train.warp.txt", index_col=0, header = 0, sep="\t")
df_test = pd.read_csv("../data/test.txt", index_col=0, header = 0, sep="\t")
df_test_warp = pd.read_csv("../data/test.warp.txt", index_col=0, header = 0, sep="\t")

# df_train.columns = [col.replace(" ", "_").replace(";", ".") for col in df_train.columns]

df_results_warp = pd.read_csv("../data/results.dbn.warp.txt", sep="\t")
df_results_model = pd.read_csv("../data/results.dbn.txt", sep="\t")

day_plus_one = [col for col in df_dag_train.columns if "day_plus_one" in col]

s_binary = {"Gender", "Supplement"}

G = nx.DiGraph()

# linear models warp
d_response_2_model_warp = {}
for response in day_plus_one:
    features = df_dag_train_warp[response]
    significant_features = features[features > 0]
    fn_count = np.sum([(("nutrient" in _) or ("food" in _)) for _ in significant_features.index])
    if significant_features.sum() > 1 and fn_count > 0:
        model_str = "Q('" + response + "')" + " ~ " + " + ".join(["Q('" + _ + "')"  if _ not in s_binary else "C(Q('" + _ + "'))" for _ in significant_features.index]) + " + 1"
        model = smf.ols(formula=model_str, data=df_train_warp)
        res = model.fit_regularized(method="elastic_net", L1_wt=0)
        d_response_2_model_warp[response] = res
        response_clean = "t+1." + response.split(".")[-1]
        G.add_nodes_from([
            (response_clean, {"color": "orange", "shape": response.split(".")[1]})
        ])
        for ix, significant_feature in enumerate(significant_features.index):
            significant_feature_clean = significant_feature.split(".")[-1]
            if significant_feature_clean not in G:
                G.add_nodes_from([
                    (significant_feature_clean, {"color": "blue", "shape": significant_feature.split(".")[0]})
                ])
            # Get coefficient from model
            G.add_edge(significant_feature_clean, response_clean)
            G[significant_feature_clean][response_clean]["color"] = "green" if res.params[ix + 1] > 0 else "red"

nx.write_gml(G, "network.gml")

res.params

res.model.exog_names

res.model.exog_names

G.add_nodes_from([
    ("test", {"color": "orange"})
    ("test2", {"color": "blue"}),
])

G.add_edge("test", "test2")
G["test"]["test2"]["color"] = "green"


nx.write_gml(G, "network.gml")

d_response_2_model_warp
